Pass along the "proceed_to" URL
This change adds an input parameter, "proceed_to", to the authorizations
controller. The goal here is to allow end users to specify where they'd
like to end up once the OAuth exchange is finished with Slack. This is
will be exposed to the end user, so the decision is totally up to them.
In most cases, I imagine this will be some sort of "Login" or "Session"
controller.