Rename slack_sign_in concern to authorization
This change renames the "slack_sign_in" method in the SignInResult
concern to be named "slack_authorization" instead. While I really liked
the resulting end-user code with the previous name, it unfortunately
conflicted with the gem's mounted routes :(. This was the best I could
come up with, but I'm going to keep thinking on it!